Set a standard password (user will be forced to update on first use)
We used to have a non-protected wifi that acted as a captive portal, forcing the user to set a password. But a lot of wireless clients have many problems switching between open and protected wifi networks with the same SSID. Therefore we now use a standard password. The user is still forced to update it before getting access to the Internet.